tvm-tutorial/requirements.txt
# TVM Tutorial Requirements

# Core dependencies
apache-tvm>=0.14.0
torch>=2.0.0
numpy>=1.20.0

# Optional but recommended
onnx>=1.12.0
onnxruntime>=1.12.0

# For visualization (optional)
matplotlib>=3.5.0
